refactor: change the output method of the output view and data type of percentage of return
1. Change the output method of the output view.
- Use printf to get rid of the number printing method using E.
- Change the constant name to put information regarding what they print.
2. Change the datatype of percentage of return from double to float.
- Float use less memory than double and percentage of return is hard to have big number so change the data type.

+++ src/main/java/lotto/model/LottoCalculate.java
@@ -51,7 +51,7 @@ public class LottoCalculate {
         return prizeNumber.prizeMoney;
     }
 
-    double calculatePercentageOfReturn(double money, double prizeMoneySum) {
-        return Math.round(prizeMoneySum / money * 1000) / 10.0;
+    float calculatePercentageOfReturn(double money, double prizeMoneySum) {
+        return (float) (Math.round(prizeMoneySum / money * 1000) / 10.0);
     }
 }
